Refactoring: Extract ui-related util methods into new class
